requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.3
django-allauth==65.15.0
django-celery-beat==2.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
